# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-admin API文档",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号相关"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "[body]",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/R.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "value": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/devops/app": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "添加应用",
                "parameters": [
                    {
                        "description": "应用信息",
                        "name": "[body]",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AppUpsertReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/R.Result"
                        }
                    }
                }
            }
        },
        "/devops/app/:id": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "更新应用",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "应用ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "应用信息",
                        "name": "[body]",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AppUpsertReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/R.Result"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "删除应用",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "应用ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/R.Result"
                        }
                    }
                }
            }
        },
        "/devops/app/download": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "下载部署包",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件Key路径",
                        "name": "key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "fileName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/R.Result"
                        }
                    }
                }
            }
        },
        "/devops/app/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "查询应用列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "按照名称模糊搜索",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页查询条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/R.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "value": {
                                            "$ref": "#/definitions/res.PageableData-dvmodel_App"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/devops/app/upload": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用管理"
                ],
                "summary": "上传部署包",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件信息",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/R.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "value": {
                                            "$ref": "#/definitions/res.UploadRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/terminal/ws": {
            "get": {
                "tags": [
                    "Web Shell"
                ],
                "summary": "websocket连接",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pty的行数",
                        "name": "rows",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pty的列数",
                        "name": "cols",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Web Shell连接使用的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "R.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "dvmodel.App": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "model.Dept": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "sortNum": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
...
        "model.Job": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sortNum": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
...
        "model.Menu": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "display": {
                    "type": "integer"
                },
                "external": {
                    "type": "integer"
                },
                "externalWay": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "sortNum": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Role": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menuList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "sortNum": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "cellphone": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "dept": {
                    "$ref": "#/definitions/model.Dept"
                },
                "deptId": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jobList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Job"
                    }
                },
                "menuList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "roleList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "root": {
                    "type": "boolean"
                },
                "sex": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
...
        "req.AppUpsertReq": {
            "type": "object",
            "required": [
                "fileName",
                "key",
                "name",
                "port",
                "version"
            ],
            "properties": {
                "fileName": {
                    "type": "string",
                    "maxLength": 100
                },
                "key": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 50
                },
                "port": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1
                },
                "version": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
...
        "req.LoginReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
...
        "res.PageableData-dvmodel_App": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dvmodel.App"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "res.UploadRes": {
            "type": "object",
            "properties": {
                "fileName": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "tmpKey": {
                    "type": "string"
                }
            }
        }
    }
}